backend-disabled/requirements.txt
fastapi
uvicorn[standard]
firebase-admin
python-decouple
